Break eval into more submodules
